ai-briefing-app/requirements.txt
fastapi
uvicorn[standard]
jinja2
supabase
python-dotenv
feedparser
requests
beautifulsoup4
readability-lxml
openai
pydantic
python-multipart
pyyaml
praw
